LRCI-4138 Create BaseJobLiferayEventHandler to share only specific methods
