+++ workspaces/liferay-jethr0-workspace/client-extensions/liferay-jethr0-etc-spring-boot/src/main/java/com/liferay/jethr0/event/liferay/LiferayEventHandlerFactory.java
@@ -41,15 +41,15 @@ public class LiferayEventHandlerFactory extends BaseEventHandlerFactory {
 				"objectActionTriggerKey");
 
 			if (objectActionTriggerKey.equals("onAfterDelete")) {
-				return new JobDeleteLiferayEventHandler(
+				return new DeleteJobLiferayEventHandler(
 					eventHandlerContext, messageJSONObject);
 			}
 			else if (objectActionTriggerKey.equals("onAfterUpdate")) {
-				return new JobUpdateLiferayEventHandler(
+				return new UpdateJobLiferayEventHandler(
 					eventHandlerContext, messageJSONObject);
 			}
 
-			return new JobAddLiferayEventHandler(
+			return new AddJobLiferayEventHandler(
 				eventHandlerContext, messageJSONObject);
 		}
